requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
openai>=1.10.0,<2.0.0

# RAG Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
chromadb==0.4.22
celery==5.3.4
redis==5.0.1

# Document Processing
pypdf2==3.0.1
python-multipart==0.0.6
Pillow==10.1.0

# Vector Database
sentence-transformers==2.2.2
